pyproject.toml
[project]
name = "enso"
version = "0.1.0"
description = "Engagement Network Scanning Orchestrator for penetration testing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "dnspython>=2.4.0",
    "netifaces>=0.11.0",
    "pytenable>=1.4.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
enso = "enso.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
